Better to support large log file when -D_FILE_OFFSET_BITS=64 on 32-bit host

+++ muduo/base/LogFile.h
@@ -19,7 +19,7 @@ class LogFile : boost::noncopyable
 {
  public:
   LogFile(const string& basename,
-          size_t rollSize,
+          off_t rollSize,
           bool threadSafe = true,
           int flushInterval = 3,
           int checkEveryN = 1024);
@@ -35,7 +35,7 @@ class LogFile : boost::noncopyable
   static string getLogFileName(const string& basename, time_t* now);
 
   const string basename_;
-  const size_t rollSize_;
+  const off_t rollSize_;
   const int flushInterval_;
   const int checkEveryN_;
